Visualization Of Process
This is a PLC program for a reactor tank. The reactor tank also has a Churner or Mixer or agitator. The tank has a lid which can be opened or closed with the help of open close buttons provided. The Motor can start automatically if the water level reaches a certain set point (which can be adjusted by the user). The program also has a motor which can be set or reset with the help of push buttons. The motor is used to drive the agitator.